requirements.txt
praw           
psycopg2-binary 
python-dotenv  
apscheduler  




streamlit          
pandas            
scikit-learn       
sentence-transformers 
torch              
ollama
redis
pydantic